# docs/zooearly-ai-openapi.yaml
openapi: 3.0.3

info:
  title: ZooEarly AI API
  version: 1.0.0
  description: |
    쥬얼리 (ZooEarly) — AI API 명세서 (v1.0 · 2026-08-21)

    React Native 앱 ↔ API Gateway ↔ FastAPI Inference Server(STT / LLM / TTS → OpenAI API)

    - 게이트웨이는 릴레이 전용: 요청 검증 + 전달 + 에러 통일만 한다. DB 없음, 무상태.
    - 상태(대화 이력, 목표 문장, 시나리오)는 전부 앱이 갖고 있으며 매 요청에 실어 보낸다.
    - FastAPI 내부 경로는 접두사만 다르다: `/api/v1/ai/chat` → `/ai/chat`
    - 어떤 에러도 아이 화면에 "오류"로 표시하지 않는다. 전부 "괜찮아, 다시 해볼까?"로 폴백한다.

servers:
  - url: https://zooearly.app/api/v1
    description: Production Gateway

tags:
  - name: AI
    description: STT / LLM / TTS 추론 릴레이

paths:
  /ai/chat:
    post:
      tags: [AI]
      operationId: chat
      summary: 음성 대화 (STT → LLM → TTS 통합 파이프라인)
      description: |
        아이의 음성을 올리면 서버가 STT → LLM → TTS를 한 번에 처리하고
        텍스트와 음성(base64 mp3)을 함께 돌려준다. 왕복 1회로 지연 최소화.

        **설계 계약**
        - STT가 못 알아들어도 `200`이다. `userText: null` + 되묻기 `aiText`가 온다.
        - 응답을 받은 앱은 `history`에 user/assistant 두 턴을 추가한다 (`userText`가 null이면 user 턴 생략).
        - 오디오 원본은 저장하지 않는다. 응답 헤더 `X-Audio-Retention: none`.
        - `aiText`는 한국어교육과정 1~2급 어휘로 생성된다 (FastAPI 프롬프트 담당).
        - 타임아웃: Gateway → FastAPI 30초 (3단 파이프라인이라 다른 엔드포인트보다 길다).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio, scenario, history]
              properties:
                audio:
                  type: string
                  format: binary
                  description: 아이의 발화. m4a/wav/webm, 최대 10MB · 60초
                scenario:
                  $ref: '#/components/schemas/Scenario'
                history:
                  type: string
                  description: |
                    지금까지의 대화를 담은 JSON 문자열. 없으면 `"[]"`.
                    최근 10턴까지만 보낸다 (앱이 자른다).
                    형식: `[{"role":"user"|"assistant","content":"..."}]`
                  example: '[{"role":"assistant","content":"불고기 많이 줄까?"}]'
                nativeLanguage:
                  $ref: '#/components/schemas/NativeLanguage'
      responses:
        '200':
          description: 대화 성공 (STT 인식 실패 포함 — userText가 null일 뿐 에러가 아니다)
          headers:
            X-Audio-Retention:
              schema: { type: string, enum: [none] }
              description: 오디오 원본 미저장 보증
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChatResult'
              example:
                success: true
                data:
                  userText: "네, 많이 주세요."
                  aiText: "그래, 많이 줄게! 맛있게 먹어."
                  audio: { data: "SUQzBAAA...", format: "mp3" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422': { $ref: '#/components/responses/SttFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502': { $ref: '#/components/responses/AiServerError' }
        '504': { $ref: '#/components/responses/AiTimeout' }

  /ai/stt:
    post:
      tags: [AI]
      operationId: stt
      summary: 음성 → 텍스트
      description: |
        음성만 텍스트로 바꾼다. SPEAK/SHADOW 스텝처럼 LLM 응답이 필요 없는 화면에서 쓴다.
        인식 결과와 목표 문장의 매칭 판정은 앱 로컬에서 한다.

        **인식 실패는 에러가 아니다** — `text: null`로 `200`이 내려간다.
        `422 STT_FAILED`는 STT 엔진 자체가 죽었을 때만 쓴다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                  description: m4a/wav/webm, 최대 10MB · 60초
                language:
                  type: string
                  default: ko-KR
                  description: BCP-47 언어 코드
      responses:
        '200':
          description: 변환 성공 (인식 실패 포함)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SttResult'
              example:
                success: true
                data: { text: "많이 주세요", confidence: 0.94 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '422': { $ref: '#/components/responses/SttFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502': { $ref: '#/components/responses/AiServerError' }
        '504': { $ref: '#/components/responses/AiTimeout' }

  /ai/tts:
    post:
      tags: [AI]
      operationId: tts
      summary: 텍스트 → 음성
      description: |
        텍스트를 음성(base64 mp3)으로 바꾼다. DIALOGUE 🔊 버튼, LISTEN 스텝에서 쓴다.

        - `voice`는 OpenAI 보이스 ID가 아니라 역할 enum이다. 실제 보이스 매핑은 FastAPI 설정에 둔다.
        - 같은 문장의 TTS 결과는 앱이 로컬 캐시한다 — 매번 서버에 묻지 않는다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  maxLength: 200
                  description: 읽을 문장
                  example: "불고기 많이 줄까?"
                voice:
                  type: string
                  enum: [TEACHER, FRIEND]
                  default: TEACHER
                  description: 역할 enum. 실제 보이스 매핑은 FastAPI 담당
                speed:
                  type: number
                  minimum: 0.5
                  maximum: 1.5
                  default: 0.9
                  description: 재생 속도. 아동용 기본 느리게
      responses:
        '200':
          description: 합성 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [audio]
                        properties:
                          audio:
                            $ref: '#/components/schemas/AudioPayload'
              example:
                success: true
                data:
                  audio: { data: "SUQzBAAA...", format: "mp3" }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502': { $ref: '#/components/responses/AiServerError' }
        '504': { $ref: '#/components/responses/AiTimeout' }

  /ai/feedback:
    post:
      tags: [AI]
      operationId: feedback
      summary: 발화 피드백 생성
      description: |
        아이의 발화(STT 결과 텍스트)와 목표 문장을 주면 FEEDBACK 스텝에 그릴 피드백 객체를 생성한다.
        음성이 아니라 텍스트를 받는다 — STT는 `/ai/stt`에서 이미 끝났다.

        **설계 계약**
        - `recognizedText: null`도 유효한 요청이다 → 격려 피드백 생성 (`understood/matched: false`).
        - `title`에 "틀렸어요"류 문구 금지.
        - `matched` 판정 활용(별점 등)은 앱 로컬 책임. 서버는 판정 재료만 준다.
        - 발화 기록은 서버에 남지 않는다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetSentence, recognizedText]
              properties:
                targetSentence:
                  type: string
                  description: 목표 문장 (앱 번들의 스텝 데이터)
                  example: "많이 주세요."
                recognizedText:
                  type: string
                  nullable: true
                  description: STT 결과. 인식 실패면 null
                  example: "많이 주세여"
                scenario:
                  $ref: '#/components/schemas/Scenario'
                nativeLanguage:
                  $ref: '#/components/schemas/NativeLanguage'
      responses:
        '200':
          description: 피드백 생성 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FeedbackResult'
              example:
                success: true
                data:
                  understood: true
                  matched: true
                  similarity: 0.92
                  title: "잘했어요!"
                  body: "무슨 뜻인지 잘 이해했어요."
                  naturalSentence: "많이 주세요."
                  naturalHint: "'주세여'보다 '주세요'가 좋아요."
                  highlightWords: ["주세요"]
                  translation: "Cho mình nhiều nhé."
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502': { $ref: '#/components/responses/AiServerError' }
        '504': { $ref: '#/components/responses/AiTimeout' }

components:
  schemas:
    Scenario:
      type: string
      enum: [ARRIVAL, CLASS, LUNCH, DISMISSAL]
      description: LLM 프롬프트 컨텍스트용 힌트. 앱 로컬의 시나리오 코드와 동일
      example: LUNCH

    NativeLanguage:
      type: string
      enum: [KOREAN, CHINESE, VIETNAMESE]
      default: KOREAN
      description: 피드백·번역 생성 언어. KOREAN이면 번역 생략
      example: VIETNAMESE

    AudioPayload:
      type: object
      required: [data, format]
      description: 서버 → 앱 오디오. chat과 tts가 동일 구조를 쓴다 — 앱의 재생 코드는 하나면 된다
      properties:
        data:
          type: string
          format: byte
          description: base64 인코딩된 mp3 바이너리
        format:
          type: string
          enum: [mp3]

    ChatResult:
      type: object
      required: [aiText, audio]
      properties:
        userText:
          type: string
          nullable: true
          description: STT 결과. 못 알아들으면 null (에러 아님)
        aiText:
          type: string
          description: LLM이 생성한 응답 문장 (한국어교육과정 1~2급 어휘)
        audio:
          $ref: '#/components/schemas/AudioPayload'

    SttResult:
      type: object
      required: [text, confidence]
      properties:
        text:
          type: string
          nullable: true
          description: 인식 결과. 못 알아들으면 null (에러 아님)
        confidence:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
          description: 엔진이 안 주면 null. 화면 표시 금지

    FeedbackResult:
      type: object
      required: [understood, matched, similarity, title, highlightWords]
      properties:
        understood:
          type: boolean
          description: 의미가 통했는가 (아이콘 결정)
        matched:
          type: boolean
          description: 목표 문장과 통했는가 (별점 판정용)
        similarity:
          type: number
          minimum: 0
          maximum: 1
          description: 화면 표시 금지. 내부 판정용
        title:
          type: string
          description: 배너 제목. "틀렸어요"류 문구 금지
        body:
          type: string
          nullable: true
          description: 배너 본문
        naturalSentence:
          type: string
          nullable: true
          description: 더 자연스러운 표현
        naturalHint:
          type: string
          nullable: true
          description: 설명문. 불필요하면 null
        highlightWords:
          type: array
          items: { type: string }
          description: 밑줄 칠 어절. 없으면 []
        translation:
          type: string
          nullable: true
          description: naturalSentence의 모국어 번역. KOREAN이면 null

    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object

    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_PARAMETER
                - UNSUPPORTED_AUDIO_FORMAT
                - AUDIO_TOO_LARGE
                - PAYLOAD_TOO_LARGE
                - STT_FAILED
                - RATE_LIMITED
                - AI_SERVER_ERROR
                - AI_TIMEOUT
                - INTERNAL_ERROR
            message:
              type: string
              description: 개발자용. 아이 화면에 띄우지 않는다
            field:
              type: string
              nullable: true
              description: 문제가 된 파라미터명. 검증 에러에만 채워진다

  responses:
    BadRequest:
      description: "검증 실패 (Gateway) — INVALID_PARAMETER / UNSUPPORTED_AUDIO_FORMAT / AUDIO_TOO_LARGE"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            success: false
            error: { code: INVALID_PARAMETER, message: "audio 필드가 없습니다.", field: audio }
    PayloadTooLarge:
      description: "요청 전체 용량 초과 (Gateway) — PAYLOAD_TOO_LARGE"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            success: false
            error: { code: PAYLOAD_TOO_LARGE, message: "요청이 너무 큽니다.", field: null }
    SttFailed:
      description: "STT 엔진 자체 실패 (FastAPI) — 인식 실패(null 반환)와 다르다"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            success: false
            error: { code: STT_FAILED, message: "STT 엔진 오류.", field: null }
    RateLimited:
      description: "OpenAI API 쿼터 초과 (FastAPI) — RATE_LIMITED"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            success: false
            error: { code: RATE_LIMITED, message: "요청이 너무 많습니다.", field: null }
    AiServerError:
      description: "FastAPI 5xx 또는 연결 불가 (Gateway 변환) — AI_SERVER_ERROR"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            success: false
            error: { code: AI_SERVER_ERROR, message: "추론 서버가 응답하지 않습니다.", field: null }
    AiTimeout:
      description: "타임아웃 초과 (Gateway) — chat 30s, 나머지 15s"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            success: false
            error: { code: AI_TIMEOUT, message: "추론 시간이 초과되었습니다.", field: null }